Move testing methods to TestCryptoTrader class
