+++ src/cryptotrader.py
@@ -43,7 +43,7 @@ class CryptoTrader:
     """
     def __init__(self, initial_balance, trade_interval, run_time, model_class, 
                  run_time_unit='h', product_id='BTC', buy_threshold=0.02, sell_threshold=0.02,
-                 order_p=0.1, confidence_threshold=0.80, slippage_p=0.005, fees_p=0.005, indicators=True, test=False, verbose=False):
+                 order_p=0.1, confidence_threshold=0.80, slippage_p=0.005, fees_p=0.005, indicators=True, verbose=False):
         # Initialize trading parameters
         self.product_id = product_id                        # Crypto symbol
         self.initial_balance = initial_balance              # Starting balance to trade
@@ -74,7 +74,7 @@ class CryptoTrader:
         # Initialize tracking parameters
         self.order_log = pd.DataFrame()
         self.verbose = verbose
-        self.test = test
+        self.test = False
 
     # Initialize time parameters
     def initialize_time(self):
@@ -243,50 +243,4 @@ class CryptoTrader:
             current_time = pd.Timestamp.now()   # Update the current time
             #TODO: Account for time it takes to trade?
     
-    # Train and run
-    def test_train_run(self, test_train_data, batch_size=32, epochs=10, seq_length=10):
-        self.test_train(test_train_data=test_train_data, 
-                        batch_size=batch_size, 
-                        epochs=epochs, 
-                        seq_length=seq_length)
-        self.test_run()
 
-    # Train model with test data
-    def test_train(self, test_train_data, batch_size=32, epochs=10, seq_length=10):
-        if not self.test: pass  # Pass if not in test mode
-
-        self.concat_indicators(test_train_data) # Concat_indicators
-
-        self.model.train(data=test_train_data, batch_size=batch_size,epochs=epochs, seq_length=seq_length)
-
-    # Run live trading loop with test data
-    #TODO: Figure this out
-    def test_run(self, test_data):
-        if not self.test: pass  # Pass if not in test mode
-        
-        self.initialize_time()  # Initialize time parameters
-
-        # Initialize the current time and end time
-        current_time = self.initial_time
-        end_time = self.end_time
-
-        for i, row in test_data.iterrows():
-            # Format data as a dataframe
-            data = pd.DataFrame([row], columns=test_data.columns)
-
-            self.concat_indicators(data)  # Add indicators to the data
-            self.update_model(data)  # Update the model
-            predicted_price, confidence = self.predict()  # Predicted price
-            current_price = data['close'][0]  # Current price
-            order, balance = self.get_order_amount()  # Order amount
-            trade_decision = self.make_trade_decision(predicted_price=predicted_price,  # Make a trade decision
-                                                       current_price=current_price,
-                                                       confidence=confidence,
-                                                       order=order)
-            self.execute_trade(trade_decision, order)  # Execute trade
-            self.log_order(trade_decision, order, predicted_price, confidence, current_price, balance)  # Log order
-            self.print_status()  # Print the current status
-
-            current_time += pd.Timedelta(seconds=self.trade_interval)
-            if current_time >= end_time:
-                break
